change!: `handshake(…)` is now generalized to support `git-receive-pack` as well.
Note that `fetch::handshake()` is still present, selecting the correct
service like before, but most of the `fetch::Ref` parsing is now handled
by `handshake::Ref` in it's more general location.

There is still `fetch::refs(…)` which just invokes a V2 ls-refs command
(without that being generalized) as `git-receive-pack` just supports
V1 at the moment, and making the `fetch::LsRefs` Command more general
would mean that every command has to be general, at is doesn't matter
anymore if it's used for Fetch or Push.

As long as git won't support V2 for pushes, which it never might,
there is no need to introduce this breakage even though it
should probably be done before going 1.0.
