+++ git-protocol/src/fetch/delegate.rs
@@ -7,7 +7,8 @@ use std::{
 use bstr::BString;
 use git_transport::client::Capabilities;
 
-use crate::fetch::{Arguments, Ref, Response};
+use crate::fetch::{Arguments, Response};
+use crate::handshake::Ref;
 
 /// Defines what to do next after certain [`Delegate`] operations.
 #[derive(PartialEq, Eq, Debug, Hash, Ord, PartialOrd, Clone, Copy)]
@@ -194,7 +195,8 @@ mod blocking_io {
 
     use git_features::progress::Progress;
 
-    use crate::fetch::{DelegateBlocking, Ref, Response};
+    use crate::fetch::{DelegateBlocking, Response};
+    use crate::handshake::Ref;
 
     /// The protocol delegate is the bare minimal interface needed to fully control the [`fetch`][crate::fetch()] operation.
     ///
@@ -254,7 +256,8 @@ mod async_io {
     use futures_io::AsyncBufRead;
     use git_features::progress::Progress;
 
-    use crate::fetch::{DelegateBlocking, Ref, Response};
+    use crate::fetch::{DelegateBlocking, Response};
+    use crate::handshake::Ref;
 
     /// The protocol delegate is the bare minimal interface needed to fully control the [`fetch`][crate::fetch()] operation.
     ///
