+++ git-protocol/src/fetch/error.rs
@@ -2,7 +2,8 @@ use std::io;
 
 use git_transport::client;
 
-use crate::fetch::{handshake, refs, response};
+use crate::fetch::{refs, response};
+use crate::handshake;
 
 /// The error used in [`fetch()`][crate::fetch()].
 #[derive(Debug, thiserror::Error)]
